Upgrade to EKS 1.15 + eksctl 0.16.0 + AWS CLI 2.07 + CA 1.15.5 (#46)
Bump all version to 1.15
Bump AWS CLI to v2.07
OnDemand Nodegroup does now run in Manage Nodegroups instead of unmanaged, that way we can explain the differences between those with Unmanaged, and with EKS Fargate.
Change in the front page to the EKS workshop
Change the order of the installation of the metric server to accommodate for installation of kube-ops-report at early stages.
Aside from Kube-ops-view, kube-resource-report is now installed to show cost allocation and savings coming from EC2 Spot nodes
Change of the allocation strategy to Capacity-optimized
Changed some of the message to highlight capacity-optimized benefits
Added extra information and changes to avoid references to Spot Instance termination handler. Using AWS Node Termination Handler instead.
Jenkins change to capacity optimized is still manual, but there is a reference to the configuration that should be used instead, this probably should change in the future.